chore: archive /developers/logo-concepts to _drafts; add placeholder
